add support for more sophisticated backups
this allows me to backup more than just the $HOME of my user on different
machines. It still uses the same layout and uses the existing FS in /backup as
a starting point of what to back up. However excludes are now read from
~/.backup/excludes and the ~/.backup folder can also include a file for each
host which a destination to backup per line.